hiberman: Get rid of the giant allocation
The idea behind performing a giant allocation during suspend
to hibernate was to force some memory to be swapped to zram,
and hence get compressed. Recently support for per process
memory reclaim was added, which pushes all eligible memory
into zram, therefore the giant allocation isn't needed
anymore.

Remove the giant allocation and stop reporting some related
metrics:

- Platform.Hibernate.MemoryAndSwapAvailable
- Platform.Hibernate.AdditionalMemoryNeeded
- Platform.Hibernate.ExcessMemoryAvailable

The calculation of the last two metrics was based on the
incorrect assumption that zram/swap can be used for the
hibernate image. Also they were recorded before the giant
allocation, which could change the values that were just
recorded. In any case it is not known exactly how much
memory is needed for the hibernate image before it is
created.

Add a new check to abort hibernation if it is very unlikely
that enough memory is available for storing the hibernate
image. This check takes an optimization into account that
allows the kernel to reduce the image size by up to 35%.
The hibernation is aborted if less than 65% of half of the
RAM is free. This is a lenient check, it can result in
false positives (the actual hibernate image is larger than
32.5% of the total RAM), but should rarely yield false
negatives. In case of a false positive the creation of the
snapshot will fail with -ENOMEM, which is explicitly
handled by hiberman.

The check for free memory is done after thickening the
hiberimage. The reason is that the page cache needs to be
dropped before the check is performed, and thickening
would re-populate the cache. As a result the hiberimage
is thickened even if the hibernation is bound to fail.
Thickening only writes about 1% of the hiberimage size,
so compared with an actual hibernation its impact on
storage wear is minor.

BUG=b:321095796, b:237007188
TEST=tast run -var cycles=3 -var simulateMemPressureMB=2500 ${DUT} power.Hibernate
  => chrome://histograms/image.size is about the same with
      and without this change

Change-Id: Ia3a18e8eecf0a077cc98796c23a6b2b29df562f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5226561
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>