dns-proxy: Fix write fd watcher in DoH client
DoH curl client was incorrectly implemented to watch write FDs using
WatchReadable method. The issue was not detected as the the read socket
will eventually trigger the curl API. The curl uprev to 8.5.0 resulted
in the read socket to not be triggered early anymore, resulting slow
first DoH query.

As the write FDs are used mostly on the initial TCP handshake,
subsequent DoH queries are not affected by the bug.

BUG=b:318765638
TEST=tast run <DUT> network.DNSProxy.arc_doh_always_on

Change-Id: I5e53db69135633dc4fe71cd0732ab0e73f72c41b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5216505
Tested-by: Jason Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>