camera: tools: mctk: Ensure remapped entities are found by name
When performing a merge operation, mctk first checks that all entities
in the source graph have a corresponding entity in the target graph.

Until now, entities whose ID had a remapping entry, but that weren't
found in the target graph by the name mentioned in the remapping entry,
received a second chance by falling back to looking them up by their
numerical ID.

This second chance turned out to be confusing when running tests,
resulting in tests that succeeded because an entity was found by ID, but
not by their remap name. For tests it is more useful to have mctk error
out if the remap operation fails halfway through - that is, if an entity
ID is to be remapped to a name, but that name is not found among the
entities in the target graph.

No more fallback to numerical ID lookup once the name lookup has failed.

Entities whose ID is not listed in the remap table continue to be looked
up by ID.

BUG=b:321163303
TEST=Created merge input files with remap entries that matched, did not match, or were absent for one ID.
     As expected, merging succeeded, failed, and succeeded.

Change-Id: I94ba9d8611703124cee79a3dff5f04d32ea61311
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5215555
Tested-by: Max Staudt <mstaudt@chromium.org>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>
Commit-Queue: Max Staudt <mstaudt@chromium.org>