biod: cryptohome: Rename V2 protos back
The separation of original/V2 protos are only for separating CLs of a
protocol change back then. The protocols are not launched yet so no
compatibility needs to be enforced. Remove the deprecated protos and
rename V2 protos back for better readability (as there's no version
handling code anywhere in the codebase).

BUG=b:298346899
TEST=build ok

Cq-Depend: chromium:5223942
Change-Id: I2d6e5d5c1e7afc9854ba3cacea9efe2dfd0d37fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5222062
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Tested-by: Howard Yang <hcyang@google.com>