trunks: Isolate the session for different sender
We need some way to distinguish the different sessions that are created
via different senders. Because we cannot load the different sessions
with the same session handle number, we need to flush/save the session
used by the other daemons.

And rename type of virtual handle to increase the code readability.

BUG=b:322943445
TEST=tast run $DUT hwsec.*

Change-Id: I4a9661d440c4d80c97ef3de6cbe2020182400bc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5224800
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: Yi Chou <yich@google.com>
Tested-by: Yi Chou <yich@google.com>