crash: KernelCollector: Fix collection of soft/hard lockups on x86
On arm64 we were successfully collecting soft and hard lockups because
the stack crawls never have any "uncertain" functions. However, on x86
uncertain functions in stack crawls are common and we need to use the
"CPU registers" to get the proper location of the soft or
hardlockup. Fix the crash collector so it handles this properly and
add a test.

NOTE: for the test, we tweak a real crash to make sure that the crash
reporter is finding the string `lkdtm_SOFTLOCKUP` from the right
place. The later instance is replaced with
`unittest_shouldnt_find_this_SOFTLOCKUP_tag`.

BUG=b:309669058
TEST=Unit tests

Change-Id: I5185c4a46d94fd67d720bfe35753be00067097ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5226567
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>