crash: KernelCollector: Fix collection of hard lockups on newer kernels
There's a minor variant in the printed message on older kernels
(ChromeOS v5.15 and older) running the downstream "buddy" detector and
newer kernels. This is because on newer kernels the print moved to
watchdog.c which has:
  #define pr_fmt(fmt) "watchdog: " fmt

Account for the variant in the regular expression we're looking for.

BUG=b:309669058
TEST=Unit tests

Disallow-Recycled-Builds: test-failures
Change-Id: I4e26da11cbf0bbb3656cd63fda6e4bb4a9ed0c90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5226941
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>