biod: cryptohome: Detect power button press during match
In the new biod flow, we should detect power button press during match
too. We can't "ignore" a press in the new flow because the attempt is
already consumed, so return a specific scan status for it. It will be
treated as error, but no UI hints will be shown by the client.

BUG=b:253991307
TEST=biod/cryptohome unittests
TEST=CQ

Change-Id: I8aa938cbafc6f35fc88b83f3128fab1f3ec37bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5222063
Reviewed-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>