hwdrm-videoproc-ta: Add memory permissions checking
This adds the TEE checks to ensure the memory references we are
accessing are of the proper type. Read for input and write for output.
Nonsecure memory because this is data to/from the nonsecure world. And
any owner because this is shared data with the nonsecure world.

BUG=None
TEST=CENCv1 H264 secure content plays on geralt

Change-Id: Ic422286726f17a5c2be980341daf7534feadb021
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5234591
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>