lorgnette: Manage firewall ports for concurrent access
Currently when a port is requested it will get closed when that client
is done with it (discovery session, open/close a scanner handle, etc).
This may cause problems if multiple clients need port access
concurrently.

Update the port management to add reference counting to the requested
ports so that a port will not get closed until all clients are finished
with that port.

BUG=b:315831201
TEST=FEATURES=test emerge-sarien lorgnette
TEST=manual test with multiple extensions and Scan App

Change-Id: I46aa2f44f26269964ba455944e32c382e27d6b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5231626
Commit-Queue: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Nathan Muggli <nmuggli@google.com>