cryptohome: Remove reset capability
As we decided that all factors that can unlock USS will be able to reset
LE credentials (wrong attempts and expiration), we don't need the reset
capabiltiy anymore. Note that none of the factors are set to only
resetting attempts and not expiration now, so this has 0 side effect.

BUG=b:314085264
TEST=cryptohome unittests

Change-Id: I50529f10aebf81d7c0053b0961ea1aaa31b3c086
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5236435
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>