patchpanel: Fix Bruschetta firewall filter
This patch first reverts crrev/c/5029373, which was totally wrong - the
original original patch crrev/c/4933740 was doing the correct thing and
crrev/c/5029373 was doing the opposite thing from what it claims in the
description, and causes DROPing everything from host to Bruschetta.

Instead, the root cause of the problem crrev/c/5029373 trying to solve
(ping6 from Bruschetta to the network did not work) was because there
had never been an FORWARD ESTABLISHED accept rule for IPv6. It did not
causes problem for other VMs because we simply accepted everything
from/to the VM, but is causing problem for Bruschetta. This patch also
fixes this problem.

BUG=b:322255158
BUG=b:279994478
TEST=unit
TEST=tast run $(DUT) network.BruschettaConnectivity.*
TEST=manual: verify Bruschetta can ping link local address of host

Change-Id: I87646bc4587dcca07067c72ddf7a0a497ac035e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5233390
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>