diagnostics: Only enable emmc_lifetime routine on models with eMMC
Currently HealthD depends on the existence of `mmc` utility program to
determine the availability of emmc_lifetime routine. However, `mmc` is
not just available on models with an eMMC drive, which makes
GetAvailableRoutines return wrong results.

Limit the routine to support the devices with "EMMC" configured in
cros-config `/hardware-properties storage-type`. That being said, the
limit is temporarily relaxed to allow devices with `storage-type`
missing or "STORAGE_TYPE_UNKNOWN" as well since many legacy eMMC models
are currently configured this way.

In addition, remove MmcSupported from SystemConfig as it is no longer
used. Unit tests for UFS in CrosHealthdDiagnosticsServiceTest are also
affected by this change as part of refactoring.

BUG=b:307882873, b:305123396
TEST=FEATURES=test emerge-$BOARD diagnostics
TEST=(eMMC dut) cros-health-tool diag get_routines → "emmc_lifetime" in the result
TEST=(eMMC dut) cros-health-tool diag emmc_lifetime → pass
TEST=(eMMC dut) tast run $dut health.Diagnostics*.emmc_lifetime → pass
TEST=(bridged-eMMC/NVMe/UFS dut) cros-health-tool diag get_routines → "emmc_lifetime" not in the result
TEST=(bridged-eMMC/NVMe/UFS dut) cros-health-tool diag emmc_lifetime → unsupported
TEST=(bridged-eMMC/NVMe/UFS dut) tast run $dut health.Diagnostics*.emmc_lifetime → skipped

Change-Id: Ieba2a63e99726d8bab05383644f1e74805f07b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5233250
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Denny Huang <dennyh@google.com>
Commit-Queue: Denny Huang <dennyh@google.com>