shill: Moving unknown retries left count to system_api
When the modem does not return a valid retries left count for given
lock type, shill uses count of 999 to represent unknown value.
But Chrome currently can not distinguish between real and unknown
count. Hence moving this constant to system_api so that it is
visible to chrome also and UI notification could be updated
accordingly.

BUG=b:317250814
TEST=Tested on Yavilla

Change-Id: I5831d5ed15cb9c8fd0706a0ace46c375eaa246fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5239207
Commit-Queue: Ujjwal Pande <ujjwalpande@google.com>
Tested-by: Ujjwal Pande <ujjwalpande@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>