cros-disks: clean up incorrect qemu test filters
The gtest filtering logic has been incorrect since the settings were
moved from the ebuild to the GN file and caused these tests to be run,
just not the same way as non-qemu.  The filter patterns are a bit
confusing, but the end result is that the tests are run as root under
qemu and as non-root under non-qemu.  This can be seen with the
--gtest_list_tests option.

Non-qemu:

$ ./disks_testrunner --gtest_list_tests --gtest_filter='*RunAsRoot*'
PlatformTest.
  RunAsRootUnmount
  RunAsRootUnmountForce
  RunAsRootCleanUpStaleMountPoints
SystemMounterTest.
  RunAsRootMount
  RunAsRootMountWithNonexistentSourcePath
  RunAsRootMountWithNonexistentTargetPath
  RunAsRootMountWithNonexistentFilesystemType

$ ./disks_testrunner --gtest_list_tests --gtest_filter='-*RunAsRoot*' | wc -l
362

Qemu:

$ ./disks_testrunner --gtest_list_tests --gtest_filter='*RunAsRoot*:DiskManagerTest.*:ExternalMounterTest.*:UdevDeviceTest.*:MountInfoTest.RetrieveFromCurrentProcess:GlibProcessTest.*' | wc -l
55

$ ./disks_testrunner --gtest_list_tests --gtest_filter='-*RunAsRoot*:DiskManagerTest.*:ExternalMounterTest.*:UdevDeviceTest.*:MountInfoTest.RetrieveFromCurrentProcess:GlibProcessTest.*' | wc -l
317

While the intention was to disable these tests under qemu, the fact
that they've been enabled & passing means we should just enable them
and be done rather than fix the filter.

This broke as part of the ebuild->gn migration in CL:4138781 & CL:4136246.
The ebuild appended the common filter settings with a - & : while the gn
only used :.
> gtest_filter_qemu_common+="DiskManagerTest.*"
> ...
> gtest_filter_user_tests="-*RunAsRoot*:"
> gtest_filter_root_tests="*RunAsRoot*-"
> gtest_filter_root_tests+="${gtest_filter_qemu_common}"

BUG=b:187795247
TEST=CQ passes

Change-Id: I734014b7d46f819df989d940bfae682f29d87bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5237776
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>