shill: move the exponential backoff logic out of PortalDetector
Originally, the PortalDetector performs the exponential backoff delay
for each trial. In the future, we want to trigger the PortalDetector and
CapportClient together. So the exponential backoff delay should be also
applied to the CapportClient.

This CL splits the exponential backoff logic to the new TrialScheduler
class, and let the NetworkMonitor controls it. With this CL, the
NetworkMonitor could trigger PortalDetector and CapportClient together
easily.

BUG=b:305129516
TEST=FEATURES=test emerge-brya shill

Change-Id: I8a76bd1032a6970b028e53dff53b57efd4af965a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5237830
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>