permission_broker: Add ST1571 to WebHID allowlist
The Scriptel ST1571 is an electronic signature pad that by default
operates in what they call "EasyScript" mode.  In EasyScript mode, the
device sends the signature capture data over a keyboard emulation
interface.  The ST1571 can also operate in what they call "ProScript"
mode.  In ProScript mode, the device uses the HID interface with the
digitizer usage page.  However, when the device is in this mode, it
exposes both the keyboard usage and the digitizer usage, even though the
keyboard interface isn't used. As a result, the device is blocked for
WebHID use.

This commit simply adds the Scriptel ST1571 to the WebHID allow list.

BUG=b:293902049
TEST=Plug ST1571 into a DUT and attempt to use WebHID APIs with the
signature pad.  Verify that permission_broker's verdict for the device
is DENY.  Update permission_broker with this change and restart the DUT,
again plug in the signature pad and attempt to use WebHID APIs with the
signature pad.  Verify that permission_broker's verdict for the device
is now ALLOW.

Change-Id: Id0ff08fdc14b5e9bedcd5c12f623b62ae3086cc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5238731
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>