hiberman: lvm: Add debug logs when command execution fails
There are crash reports that indicate that hiberman couldn't
determine the volume group of the stateful partition. From the
error log it isn't clear whether the failure is due to repeated
ECMD_FAILED errors or something else. Add debug logs when the
execution of a LVM command fails to learn more about the issue.

BUG=b:302612046
TEST=emerge-${BOARD} hiberman

Change-Id: I967627043a08e93ee35191699d3b4195798c9496
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5246866
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>