crash: Moved "shared_defines" into pkg_config
In the BUILD.gn, make sure the shared_defines are all in pkg_configs,
and that the pkg_configs are included with all_dependent_configs in
libraries.

This should lower the chances of a ODR violation where one cc files
includes a header with one set of defines and another includes the
header with a different set of defines.

BUG=None
TEST=emerge-trogdor crash-reporter

Change-Id: I60ac8dd7a7ddd158101eff52d7ed37c5e1d5110e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5238737
Auto-Submit: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>