shill: portal: set validation mode from Service
This patch adds a Service getter for obtaining the initial network
validation mode when a Network connection starts. The initial validation
mode is passed to the Network in the Network::StartOptions argument.

In addition a Network setter is added to allow the Service to update the
validation mode of its attached network when a relevant property is
updated.

This patch has no behavior change. A follow-up patch will use the
validation mode setter inside Network to enable or disable network
validation and replace the equivalent logic in
Service::UpdateNetworkValidation.

BUG=b:314693271
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I13b86452749cd7b61ed40efdc4e367ba409bd776
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5267318
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>