patchpanel: Do not use StrictMock<Datapath> in QoSServiceTest
`StrictMock` is expensive to compile if the base mock class is huge.

This patch changes to `StrictMock<Datapath>` use the normal mock class
with `Times(0)` expectation. This helps reduce the compile time for
qos_service_test.cc from 15s to 10s, and also be better in readability.

BUG=None
TEST=FEATURES="test" emerge-$BOARD patchpanel

Change-Id: Ie0b96c01617ef51d97db9414f5450a1c47edebc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5239658
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>