debugd: add top-level user / daemon store size dump to report
They do not contain PII, contrary to the comment in the code
Example: https://paste.googleplex.com/6312104682848256

And that should be fine according to the privacy consultation
http://b/227038875#comment23

The only exception is that stderr may leak underlying files.
Code is changed to ignore the stderr for the user/daemon-store dump.

BUG=b:227038875
TEST=CQ

Change-Id: I0daff577560f643cdb5c6f7324cb5f1db8c6e29c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5246725
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Xiangdong Kong <xiangdongkong@google.com>
Tested-by: Daniil Lunev <dlunev@chromium.org>