cryptohome: Add error reporting to mount started
This implements the error fields added to this signal. The code
implements this by wrapping the on_done signal in all cases; this is
somewhat overkill for some of the functions but it allows the "populate
the signal" to be reused as much as possible.

BUG=b:316597268
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I666d63c228276cbe2d6107dfa8cfe8e457f5b731
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5246319
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>