dlcservice: remove dependency on "bc"
It was either this, or add a dependency in the ebuild.
But since bash is perfectly capable of doing that
calculation, use bash.

Also probably not worth having a separate function just to
invoke "stat" since the comments make it clear what goes on.

BUG=none
TEST=manually checked the expression

Change-Id: I6ad9fb971bb173ceb86d9896986504cfb5027224
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5247457
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Jae Hoon Kim <kimjae@chromium.org>