oobe_config: Check Flex config file path instead of dir
We can't delete the flex_config directory as part of cleanup post-OOBE,
since it's parent is root-owned. Given this, we should check the
presence of the actual config.json file instead before attempting to
read the file, as the directory may be a remnant of a previous
auto-enrollment.

BUG=b:316944501
TEST=cros_run_unit_tests --board reven --packages oobe_config

Change-Id: Ie5ebee9e0a1fb2d51499acdbceb6c2eace853f44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5248209
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Jackson Tadie <jacksontadie@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>