shill: change NetworkMonitor::Start() return asynchronously
The NetworkMonitor::Start() method is used to schedule a validation
trial, so the method is an asynchronous function by nature.

This CL changes the NetworkMonitor::Start() method to return the result
asynchronously by ClientNetwork::OnValidationStarted(bool). With this,
we could postpone getting the network configuration until the validation
trial is started.

BUG=b:305129516
TEST=FEATURES=test emerge-brya shill

Change-Id: Ic421e5da45a6ba71641662e1317c702f0917eed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5248490
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>