patchpanel: Implement TagSocket() API
Map the request to the corresponding bits in fwmark (network_id to
rt_table_id, vpn_policy to policy) and then call setsockopt() on the
socket().

Note that network_id does not exist now, and thus this patch just treats
the network_id in the request as an interface index, so that the
functionality of this API can be tested now. We will change it to the
real implementation after network_id is ready.

BUG=b:322083502
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=Verified the routing semantics with TagSocket when a VPN is
  connected. (tests for VPN lockdown mode is still WIP.)

Change-Id: I1d123fe7dedec0357295ce76517d0fbea6220cd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5248586
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>