cryptohome: Extract some of the API init in UDA to an external object
This takes some of the initialization of the internal API objects that
UserDataAuth does and pulls them out into an external object which is
shared via pointers passed into UDA. This changes doesn't attempt to
extract all of the objects which can be extracted this way, just a
smaller initial set to demonstrate the basic mechanics.

The idea is to reduce the amount of "create object, then inject all of
the dependencies with setters" behavior in testing. The handling of this
behavior is currently rather messy, especially because we have a bunch
of cases where tests that are trying to inject mocks for the lowest
layer of objects also have to override a bunch of composite objects
built on top of them. It also can introduce some weird initialization
ordering issues when the construction and overrides aren't sequenced
properly and you end up with a mix of different objects in use; the
constructor based approach forces the tests to set up the dependencies
first.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I079ec17e0aa5e489bfdf9607f2f538f67b915ed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5251734
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>