regions: Use TBR in unconfirmed region data
The 'unconfirmed region data' was generated from public documents (
experimental/generator.py), but that contains a lot of data that is not
supported by Chrome and ChromeOS, or even not Linux. However, having
such data may lead to confusion when approving a new region because they
look like 'pre-approved'.

To prevent the confusion we are now changing all unconfirmed values to
'TBR' (to be reviewed) to make it more explicit. Also re-ordered and
added more comments to nordic and latam-es-419 regions so it is easier
to figure out the right region code to use before people tries to
review/approve them without knowing the existing codes.

BUG=None
TEST=FEATURES=test sudo emerge region

Cq-Depend: chromium:5252392
Change-Id: I7054238bf4c4e1428906f84033f6a5ca199a6cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5252412
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Auto-Submit: Hung-Te Lin <hungte@chromium.org>