hwsec-utils: Fix inccorect futility flags
The Rust migration of cr50-reset script kept the behavior of single
quoting the value `+0x8` for gbb flags, however the generated
command line argument cannot be parsed correctly by the `futility`.
We tried to escape the single quote as a way to fix but it did not
work, so as long as the single quote is not neccessarily needed,
we just simply remove it in this CL to make things work as intended.

BUG=b:322736627
TEST=emerge hwsec-utils and install the gsc_reset binary onto the DUT,
verify that the gbb flags can be set to 0x8 correctly after action_e.

Change-Id: I3b7c8ffe3e7f12216191e3d56528a70aba50d3e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5252588
Reviewed-by: Sam Chen <chensa@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Wei-Shuo Su <stevesu@google.com>
Commit-Queue: Wei-Shuo Su <stevesu@google.com>