print_tools: Resolve mdns names in printer_diag
The ChromeOS configuration of libcurl directly resolves DNS names, so it
does not support connecting to names in the .local mdns space.  Before
connecting to a printer URI that contains one of these names, use the
system resolver to look it up and replace the hostname with the first IP
returned.

BUG=None
TEST=printer_diag --url=ipp://hostname.local/ipp/print

Change-Id: I4260760a4f9e1cc973b6438bc393ace5c774fd79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5249586
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>