cryptohome: Extract more UDA APIs
This pulls more of the lower level API objects out of UDA initialization
and into system APIs.

This also adds a MockSystemApis utility class to make it easier to
declare all your mocks in unit tests. Instead of every UDA test having
to replicate mock setup for everything that the non-test version does
for free, they can just use MockSystemApis instead. The change is a bit
noisy because it forces a bunch of variable name changes in the test,
but otherwise is a fairly mechanical drop-in.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Iee641cc386a39815b60ccadc3fb7c54761c23195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5254000
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>