federated: Add writev to syscall filters
This is causing crashes in tests and in the field. This likely
isn't the actual root cause of the crashes (the in-the-field
traces indicate an __assert_fail), but should give us a bit
more useful logging in understanding what exactly is failing.

Crash corp:
http://shortn/_rxciDxbllm

Example report:
http://shortn/_eHATl1SKjc

Example test failure in testhaus:
http://shortn/_yEInglWQOV

BUG=None
TEST=CQ

Change-Id: I4c0552eb25c308a68816d17fa2e45cc1c4bdc2b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5254003
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>