chromeos-config: Output error to kmsg
A partner changed the HWID of their device to an invalid value to test
a feature and ran into an issue booting up their device. It is not clear
why a cros_config failure would cause the stateful partition to not mount,
but in the meanwhile, outputting the error message directly to dmesg
should help figure out similar situations in the future more easily.

BUG=b:319545845
TEST=Manual check with a valid HWID and with an invalid one.

Change-Id: I762708ea76d990ea1bc3fd10cda632527ced785e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5255557
Tested-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Paz Zcharya <pazz@google.com>