oobe_config: Allow oobe_config DBus calls from Tast
Add "root" as an allowed caller of OobeConfigRestore DBus methods, so
that we can test these methods via Tast.

Another method for invoking DBus from Tast was suggested in b/322027482,
namely to use textexec.CommandContext to run dbus-send as the chronos
user. This has a couple benefits (like making this CL unnecessary), but
proved to be a really ugly and time-consuming implementation because
we'd need to parse the textual output of dbus-send and marshall it to a
proto. In the interest of efficiency, I prefer the using the approach of
using the existing Tast DBus utilities unless the reviewers feel
strongly.

BUG=b:322027482
TEST=tast run <device> policy.FlexConfigDbusSmoke

Change-Id: I76d643f4aa2af767be67147be00424ed89ae2fe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5255717
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Jackson Tadie <jacksontadie@google.com>
Commit-Queue: Jackson Tadie <jacksontadie@google.com>