ml_core: Only trigger DlcClient once
DlcClient is used for triggering install of the ML Core DLC.
It takes OnceCallbacks for success and failure.
Under certain conditions its possible for a DLC to initially fail, but then succeed later.
(This can happen if multiple clients try to install the same DLC, or if DLC is installed via CLI)

This change causes DlcClient to only trigger the first callback.

BUG=322244524

Change-Id: I41ff665b1d9db2184fc57ab8174090c60ec819e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5256677
Tested-by: Nikolas Bowe <nbowe@google.com>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Nikolas Bowe <nbowe@google.com>