power: Add intel_prepost_suspend command
Add the executable:
  * intel_prepost_suspend

This new command allow for performing Intel SOC related command. Usually
the tester needs to run many scripts or commands before/after test.
The executable could simplify the debugging process. We don't need to
provide detail and complex scripts to the tester again and again.
It also could avoid the tester forgetting any steps then need to
reproduce the issue again and again.

First time, we add ltr_ignore for it. Usually the suspend fail caused by
ltr could be temporarily waived. Before it got fixed, we need to keep
testing to check if other fail behind it. So the tester always need to
execute the ltr ignore command. With this executable, we don't need to
double confirm again whether the tester execute the ignore command.
The tester just need to execute the suspend with
--pre_suspend_command="intel_prepost_suspend --ltr_ignore=ISSUED_IP"

BUG=b:303771771
TEST=intel_prepost_suspend --ltr_ignore="10", check the ignore is set

Change-Id: Ic24f5eff2526234e38cdd470d1bcbe220002e378
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5256678
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Curtis Chen <curtis.chen@intel.com>
Tested-by: Curtis Chen <curtis.chen@intel.com>