shill: keep PortalDetector always alive in NetworkMonitor
Originally, we destroy the PortalDetector instance when stopping the
network validation, and create a new one when starting the validation.

This CL makes the NetworkMonitor not destroying the PortalDetector.
Instead, it only calls PortalDetector::Reset() method to reset the
internal state. With that, NetworkMonitor could access the
PortalDetector anytime without worrying about null-pointer crash.

Also, this CL renames some variables of PortalDetector for consistency.

BUG=b:305129516
TEST=FEATURES=test emerge-brya shill

Change-Id: Ia2c6af1e6f28334ece4b2b0385b06de583e5cc24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5256838
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>