login: Add retry mechanism to liveness checker
In the current implementation if a dbus message is dropped for whatever
reason, browser will be restarted after the timeout expires.
Transient dbus failures don't necessarily imply that the browser itself
is hung, so that behavior is not ideal.
As a work around add a retry mechanism. Instead of sending a single
message with 60s dbus timeout, reduce its timeout to 6s, but allow it to
be re-tried 10 times.
As a nice side effect this allows us to capture the chrome state
earlier, providing us with more information about shorter hangs.
For the time being this new behavior is limited to beta and dev channels
only.

BUG=b:303689265
TEST=pkill -SIGSTOP chrome
dbus-monitor --system interface=org.chromium.LivenessServiceInterface
and verify that messages are sent every 6s

Change-Id: I5363246fdfe87dcb4aa8e2d0cddd5e802398365e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5257297
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Kornel Dulęba <korneld@google.com>
Tested-by: Kornel Dulęba <korneld@google.com>