init: Remove cryptohome LVM in test dev mode transition
On the testimage, we would skip the full stateful clobbering to preserve
the testing tools under /usr/local, but the side effect on the LVM
devices also preserve the cryptohome related data.
And that will confus the developers and OEMs, because after the dev mode
transition, they will think the device is in a "clean state", but the
remaining cryptohome data would not let them login.

We should catch this corner case to prevent more confusion on the test
image.

BUG=b:257193446
TEST=Install the test image on the brya (LVM enabled).
     Login into an account successfully.
     Do a dev mode transition.
     The device go back to OOBE and can login into same account again.

Change-Id: I670a08bc93fe53c6bb59e6f4d352dc0b246bf24a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5256323
Tested-by: Yi Chou <yich@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Yi Chou <yich@google.com>